requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
pandas==2.1.4
django-cors-headers==4.3.1
django-ratelimit==4.0.0
gunicorn==21.2.0
